+++ crates/tests/node-tests/cache_response.rs
@@ -1,7 +1,7 @@
 use anyhow::{Result, anyhow};
 use node::{
     AcknowledgeRequest, BlindedMessage, MeltRequest, MintQuoteRequest, MintRequest,
-    utils::hash_mint_quote_request,
+    hash_mint_quote_request,
 };
 use node_tests::{init_health_client, init_node_client};
 use tonic_health::pb::{HealthCheckRequest, health_check_response::ServingStatus};
